+++ src/generator/rules/php_composer.js
@@ -1,65 +1,37 @@
-import { _, log } from 'azk';
+import { log } from 'azk';
 import { BaseRule } from 'azk/generator/rules';
-var semver = require('semver');
-
-var getVersion = function(path, content) {
-  var parsedJson;
-  try {
-    parsedJson = JSON.parse(content);
-  } catch (err) {
-    log.error('JSON.parse error [', path, ']', err.stack || err);
-  }
-
-  if (parsedJson &&
-      parsedJson.require &&
-      parsedJson.require.php) {
-    // remove garbage
-    var versionCleaned = parsedJson.require.php.replace(/[^\d\.\*]/g, "");
-    return semver.clean(versionCleaned);
-  }
-
-  return null;
-};
 
 export class Rule extends BaseRule {
   constructor(ui) {
     super(ui);
-    this.type = 'runtime';
+    this.type      = "runtime";
+    this.name      = "php_composer";
+    this.rule_name = "php_composer";
+    this.replaces  = ['php', 'node'];
+
+    // Suggest a docker image
+    // http://images.azk.io/#/php-fpm
+    this.version_rules = {
+      'php_composer-5.5' : '>=5.0.0 <5.6.0',
+      // 'php_composer-5.6' : '<5.5.9 || >=5.6.3'
+    };
   }
 
   relevantsFiles() {
     return ['composer.json'];
   }
 
-  getEvidence(path, content) {
-    var evidence = {
-      fullpath: path,
-      ruleType: 'runtime',
-      name    : 'phpcomposer',
-      ruleName: 'php56',
-      replaces: ['php', 'node']
-    };
-
-    var phpVersion = getVersion(path, content);
-    evidence.version = phpVersion;
-
-    // cant find php version, will use default php v5.5.9
-    if (!phpVersion) {
-      return evidence;
+  getVersion(content) {
+    var parsedJson, version;
+    try {
+      parsedJson = JSON.parse(content);
+    } catch (err) {
+      log.error('JSON.parse error', err.stack || err);
     }
-
-    // Suggest a docker image
-    // https://registry.hub.docker.com/u/azukiapp/php-fpm/
-    var versionRules = {
-      'php55' : '>=5.5.9 <5.6.3',
-      'php56' : '<5.5.9 || >=5.6.3'
-    };
-
-    evidence.ruleName = _.findKey(versionRules, (value) => {
-      return semver.satisfies(phpVersion, value);
-    });
-
-    return evidence;
+    if (parsedJson && parsedJson.require && parsedJson.require.php) {
+      // remove garbage
+      version = parsedJson.require.php.replace(/[^\d\.\*]/g, "");
+    }
+    return version && this.semver.clean(version);
   }
-
 }